+++ packages/svelte/CHANGELOG.md
@@ -6,6 +6,10 @@ description: All notable changes will be documented in this file.
 
 ## [Unreleased]
 
+### Added
+
+- **Field**: Add `data-required` attribute to Field.Label part
+
 ## [5.7.0] - 2025-08-28
 
 ### Added